+++ Optimizer/HostsEditorForm.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.IO;
 using System.Windows.Forms;
 
 namespace Optimizer