+++ Optimizer/Integrator.cs
@@ -1,11 +1,11 @@
-﻿using System;
+﻿using Microsoft.Win32;
+using System;
 using System.Collections.Generic;
-using System.IO;
-using Microsoft.Win32;
-using System.Windows.Forms;
 using System.Drawing;
-using System.Net;
 using System.Drawing.Imaging;
+using System.IO;
+using System.Net;
+using System.Windows.Forms;
 
 namespace Optimizer
 {