+++ Optimizer/SilentOps.cs
@@ -1,7 +1,7 @@
-﻿using System.Threading.Tasks;
-using Newtonsoft.Json;
-using System.IO;
+﻿using Newtonsoft.Json;
 using System;
+using System.IO;
+using System.Threading.Tasks;
 
 namespace Optimizer
 {